Rebase and sync with downstream repo
Rebase and sync the contents of the downstream / environment repo while
also dropping existing environments. The expectation is that this repo
will be used as a base repo from a private environment repo and patching
contents for the particular environment.

In the future an example environment/example/ may be created.