Add GitOps install automation (#7)
Because we don't have GitOps yet, we can't use all the automation
provided by that system. The initialization of a GitOps-managed
environment is GitOps itself. This script uses Ansible in conjunction
with the kubernetes.core collection to manage the deployment of GitOps
using the kustomize manifests.

Also significant updates to the README to make bootstrapping easier,
whether that is for a hub cluster, or an unmanaged cluster. More work to
document creating GitOps Applications and applying them to the hub or
unmanaged cluster will be required in the future.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>