Add base dependencies as flat files (#8)
Add the base dependencies for RHOSO with a simple GitOps Application,
without the use of anything from validated architectures. Provides the
prerequisite Operators which in documentation is done via the OpenShift
UI normally. This is a set of optional installation manifests which
could be performed by the administrator prior to running
environment-specific installation with GitOps.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>